// src/app/shared/enums.ts
export enum Mode {
  Create = 'create',
  Edit = 'edit'
}

export enum ModalProjectTitle {
  Create = 'Create Project',
  Edit = 'Edit Project'
}

export enum ModalTaskTitle {
  Create = 'Create Task',
  Edit = 'Edit Task'
}

export enum Status {
  Todo = 'todo',
  Doing = 'doing',
  Done = 'done'
}